Example error cleanup from Breseq output
This Jupyter notebook shows the code used to cleanup metagenomics data from 1000 generations of evolution of a mini-community of two species, D. vulgaris and M. maripaludis. The data was processed with sickle, scythe, and then Breseq. A csv file that includes the frequencies of all putative mutations for several timepoints in evolution was used as the input file.